Rename new_history.py to history.py
