+++ src/estimagic/optimization/tranquilo/history.py
