+++ tests/optimization/tranquilo/test_wrap_criterion.py
@@ -2,7 +2,7 @@ import itertools
 
 import numpy as np
 import pytest
-from estimagic.optimization.tranquilo.new_history import History
+from estimagic.optimization.tranquilo.history import History
 from estimagic.optimization.tranquilo.wrap_criterion import get_wrapped_criterion
 from numpy.testing import assert_array_almost_equal as aaae
